refactor(reconcilers): flatten src/reconcilers/ subdir into crate root
The `overdrive-reconcilers` crate carried its impls under a
`src/reconcilers/` subdirectory, so every submodule-qualified path read
`overdrive_reconcilers::reconcilers::<sub>::X` — the `reconcilers`
segment is pure redundancy against the crate name. Flatten it:

- Move the 7 impl modules (backend_discovery_bridge, noop_heartbeat,
  service_map_hydrator, svid_lifecycle, vm_reclamation,
  workflow_lifecycle, workload_lifecycle) up one level into `src/`.
- Fold `reconcilers/mod.rs` — the three dispatch enums
  (AnyState / AnyReconciler / AnyReconcilerView), the flat re-exports,
  and the workload_id_from_target / service_id_from_target helpers —
  into `lib.rs`. The impls' `super::` references resolve unchanged
  because `super` from a `src/*.rs` module is now the crate root.
- Rewrite importers workspace-wide:
  `overdrive_reconcilers::reconcilers::<sub>::` -> `overdrive_reconcilers::<sub>::`.
  The `overdrive_core::reconcilers` contract module (a different crate)
  is untouched.
- Fix the dst-lint gate's runtime path resolver (`reconciler_source_path`
  in xtask) that `read_to_string`s the moved files, plus stale
  doc-comment path references.

Pure module-layout change: the flat public API
(`overdrive_reconcilers::AnyReconciler`, `::WorkloadLifecycle`, ...) is
unchanged; only submodule-qualified paths shed the `::reconcilers`
segment. ADR-0086's contract-in-core split is preserved.
