+++ crates/overdrive-control-plane/tests/acceptance/bridge_emits_enqueue_evaluation_for_hydrator.rs
@@ -19,12 +19,12 @@ use std::num::NonZeroU16;
 use overdrive_core::dataplane::backend_key::Proto;
 use overdrive_core::id::{AllocationId, NodeId, ServiceId, ServiceVip, WorkloadId};
 use overdrive_core::reconcilers::Reconciler;
-use overdrive_reconcilers::reconcilers::backend_discovery_bridge::{
+use overdrive_core::reconcilers::{Action, TickContext};
+use overdrive_core::wall_clock::UnixInstant;
+use overdrive_reconcilers::backend_discovery_bridge::{
     BackendDiscoveryBridge, BackendDiscoveryBridgeState, BackendDiscoveryBridgeView,
     ProjectedListener,
 };
-use overdrive_core::reconcilers::{Action, TickContext};
-use overdrive_core::wall_clock::UnixInstant;
 use std::net::IpAddr;
 use std::time::{Duration, Instant};
