+++ crates/overdrive-control-plane/tests/acceptance/svid_lifecycle_runtime_backoff_reenqueue.rs
@@ -37,7 +37,6 @@ use overdrive_control_plane::reconciler_runtime::{ReconcilerRuntime, run_converg
 use overdrive_control_plane::{AppState, noop_heartbeat, svid_lifecycle};
 use overdrive_core::eval_broker::Evaluation;
 use overdrive_core::id::{AllocationId, NodeId, WorkloadId};
-use overdrive_reconcilers::reconcilers::svid_lifecycle::{IssueRetry, SvidLifecycle, SvidLifecycleView};
 use overdrive_core::reconcilers::{Reconciler, ReconcilerName, TargetResource};
 use overdrive_core::traits::ca::Ca;
 use overdrive_core::traits::driver::{Driver, DriverType};
@@ -46,6 +45,7 @@ use overdrive_core::traits::observation_store::{
     AllocState, AllocStatusRow, LogicalTimestamp, ObservationRow, ObservationStore,
 };
 use overdrive_core::wall_clock::UnixInstant;
+use overdrive_reconcilers::svid_lifecycle::{IssueRetry, SvidLifecycle, SvidLifecycleView};
 use overdrive_sim::adapters::ca::SimCa;
 use overdrive_sim::adapters::clock::SimClock;
 use overdrive_sim::adapters::dataplane::SimDataplane;
