+++ crates/overdrive-core/tests/acceptance/svid_lifecycle_reconcile.rs
@@ -25,16 +25,19 @@ use overdrive_core::aggregate::{Exec, Job, Node, WorkloadDriver, WorkloadKind};
 use overdrive_core::id::{
     AllocationId, ContentHash, CorrelationKey, NodeId, Region, SpiffeId, WorkloadId,
 };
-use overdrive_reconcilers::reconcilers::svid_lifecycle::{
-    IssueRetry, RunningAlloc, SvidLifecycle, SvidLifecycleState, SvidLifecycleView,
-};
 use overdrive_core::identity::HeldSvidFacts;
 use overdrive_core::reconcilers::{Action, Reconciler, TargetResource, TickContext};
-use overdrive_reconcilers::{RESTART_BACKOFF_CEILING, WorkloadLifecycle, WorkloadLifecycleState, WorkloadLifecycleView, backoff_for_attempt};
 use overdrive_core::traits::driver::Resources;
 use overdrive_core::traits::observation_store::{AllocState, AllocStatusRow, LogicalTimestamp};
 use overdrive_core::transition_reason::TerminalCondition;
 use overdrive_core::wall_clock::UnixInstant;
+use overdrive_reconcilers::svid_lifecycle::{
+    IssueRetry, RunningAlloc, SvidLifecycle, SvidLifecycleState, SvidLifecycleView,
+};
+use overdrive_reconcilers::{
+    RESTART_BACKOFF_CEILING, WorkloadLifecycle, WorkloadLifecycleState, WorkloadLifecycleView,
+    backoff_for_attempt,
+};
 use proptest::prelude::*;
 
 const NODE_RAW: &str = "local";
