+++ crates/overdrive-core/tests/backend_discovery_bridge_types.rs
@@ -30,12 +30,12 @@ use std::collections::BTreeMap;
 use std::net::Ipv4Addr;
 
 use overdrive_core::id::{CorrelationKey, NodeId, ServiceId};
-use overdrive_reconcilers::reconcilers::backend_discovery_bridge::{
+use overdrive_core::reconcilers::Action;
+use overdrive_core::traits::observation_store::{LogicalTimestamp, ServiceBackendRow};
+use overdrive_reconcilers::AnyReconciler;
+use overdrive_reconcilers::backend_discovery_bridge::{
     BackendDiscoveryBridge, BackendDiscoveryBridgeView,
 };
-use overdrive_core::reconcilers::{Action};
-use overdrive_reconcilers::{AnyReconciler};
-use overdrive_core::traits::observation_store::{LogicalTimestamp, ServiceBackendRow};
 
 /// T-BDB-VIEW-1 (ADR-0079 § D3 / § D7) — a CBOR blob written by a
 /// PRE-ADR-0079 binary, carrying a populated `last_written_fingerprint`
