+++ crates/overdrive-core/tests/service_frontend_provenance.rs
@@ -34,16 +34,16 @@ use std::time::{Duration, Instant};
 
 use overdrive_core::dataplane::backend_key::Proto;
 use overdrive_core::id::{NodeId, ServiceId, ServiceVip, SpiffeId};
-use overdrive_reconcilers::reconcilers::service_map_hydrator::{
-    ServiceProjectionError, project_service_desired,
-};
 use overdrive_core::reconcilers::{Action, Reconciler, TickContext};
-use overdrive_reconcilers::{ServiceMapHydrator, ServiceMapHydratorState, ServiceMapHydratorView};
 use overdrive_core::traits::dataplane::Backend;
 use overdrive_core::traits::observation_store::{
     ListenerRow, LogicalTimestamp, ServiceBackendRow, ServiceBackendRowLatest,
 };
 use overdrive_core::wall_clock::UnixInstant;
+use overdrive_reconcilers::service_map_hydrator::{
+    ServiceProjectionError, project_service_desired,
+};
+use overdrive_reconcilers::{ServiceMapHydrator, ServiceMapHydratorState, ServiceMapHydratorView};
 
 fn vip_v4(o: u8) -> ServiceVip {
     ServiceVip::new(IpAddr::V4(Ipv4Addr::new(10, 96, 0, o))).expect("valid IPv4 ServiceVip")
