+++ crates/overdrive-sim/src/invariants/vm_reclamation.rs
@@ -33,9 +33,9 @@ use std::path::PathBuf;
 
 use overdrive_core::cgroup::CgroupPath;
 use overdrive_core::id::{AllocationId, WorkloadId};
-use overdrive_core::reconcilers::{Action};
-use overdrive_reconcilers::{SupervisionSet, VmAllocFacts, VmReclamationState, plan_reclamation};
+use overdrive_core::reconcilers::Action;
 use overdrive_core::traits::vm_host_state::{ScopeFacts, VmHostState};
+use overdrive_reconcilers::{SupervisionSet, VmAllocFacts, VmReclamationState, plan_reclamation};
 
 use crate::adapters::clock::SimClock;
 use crate::adapters::vm_host_state::SimVmHostState;
@@ -242,7 +242,7 @@ pub async fn evaluate_vm_reclamation_converges() -> InvariantResult {
 
     let sim = SimVmHostState::new();
     // DST-controllable stand-in for the 30s sweep cadence
-    // (`overdrive_reconcilers::reconcilers::vm_reclamation::VM_RECLAMATION_SWEEP_INTERVAL`) —
+    // (`overdrive_reconcilers::vm_reclamation::VM_RECLAMATION_SWEEP_INTERVAL`) —
     // `plan_reclamation` itself consults no clock (pure over `(desired,
     // actual)` alone), so advancing the SAME ratified constant here
     // documents the simulated pacing a real `spawn_convergence_loop` tick
@@ -281,7 +281,7 @@ pub async fn evaluate_vm_reclamation_converges() -> InvariantResult {
 
     let mut converged = false;
     for _tick_n in 0..TICK_BUDGET {
-        clock.tick(overdrive_reconcilers::reconcilers::vm_reclamation::VM_RECLAMATION_SWEEP_INTERVAL);
+        clock.tick(overdrive_reconcilers::vm_reclamation::VM_RECLAMATION_SWEEP_INTERVAL);
         let host = match sim.observe().await {
             Ok(h) => h,
             Err(e) => return fail(NAME, format!("observe failed: {e}")),
